Move waitUntilFind module
