+++ client/startup/e2e.ts
@@ -7,11 +7,11 @@ import { Subscriptions, Rooms } from '../../app/models/client';
 import { Notifications } from '../../app/notifications/client';
 import { promises } from '../../app/promises/client';
 import { settings } from '../../app/settings/client';
-import { waitUntilFind } from '../../app/utils/client';
 import { IMessage } from '../../definition/IMessage';
 import { IRoom } from '../../definition/IRoom';
 import { ISubscription } from '../../definition/ISubscription';
 import { isLayoutEmbedded } from '../lib/utils/isLayoutEmbedded';
+import { waitUntilFind } from '../lib/utils/waitUntilFind';
 
 const handle = async (roomId: IRoom['_id'], keyId: string): Promise<void> => {
 	const e2eRoom = await e2e.getInstanceByRoomId(roomId);
